Move AreaHelpers to its own file
